Initial subforms implementation.
This required changes to inject, autosubmit, and the input-change-event
handling:

1. Inject can now be installed on .pat-subform elements. The injection
is triggered by the submit/clickSubmit handlers of pat-subform but uses
the configuration parser and semantics from pat-inject.
2. Inject.submitSubform performs injection on the form element but uses
the configuration stored on the subform element.
3. The pat-inject url parameter is now usable, since .pat-subform
elements usually do not have href or action attrs.
4. Autosubmit and input-change-events are now aware of subforms.

Subform submission is performed by the native submit function or by
ajaxSubmit, parts of the form that are to be excluded from the
transmission are stripped of their name during the process.

There is initial support for formaction on button in subforms.